Todas las velocidades y posiciones pasadas a float para hacer un movimiento mas preciso

+++ src/game/cmp/score.hpp
@@ -7,6 +7,6 @@ struct ScoreComponent_t : public ECS::ComponentBase_t<ScoreComponent_t> {
         : ComponentBase_t(eid)
     {}
     uint32_t score { 0 };
-    inline static uint32_t scorePosX { 0 };
-    inline static bool     scored    { false };
+    inline static float scorePosX { 0 };
+    inline static bool  scored    { false };
 };
